facial_recognition_system/requirements.txt
# Core dependencies
face-recognition>=1.3.0
opencv-python>=4.5.0
numpy>=1.19.0
dlib>=19.22.0
scipy>=1.5.0

# Web/API
flask>=2.0.0
flask-cors>=3.0.10
Pillow>=8.0.0

# Security
bcrypt>=3.2.0
PyJWT>=2.0.0

# Configuration
pyyaml>=5.4.0

# Testing
pytest>=6.0.0
pytest-cov>=2.12.0

# Utilities
tqdm>=4.62.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
requests>=2.26.0

# Optional but recommended
redis>=4.0.0
celery>=5.1.0
gunicorn>=20.1.0  # For production deployment